fix: nmea2000 output with canboat
nmea2000out configured as stdout event by streams/simple tried to
also subscribe to nmea2000JsonOut events with conversion in place,
but because the event name was not correct this has never worked.

This commit fixes the event name, writes to child process's stdin
and adds require for the conversion function.